Reworking checkpoint & managers
